requirements.txt
flask
python-dotenv
openai
pyrebase
firebase_admin